+++ peptonizer_rust/src/fetch_unipept_taxa.rs
@@ -21,15 +21,15 @@ pub fn fetch_peptides_and_filter_taxa(
 ) -> Result<String, Box<dyn std::error::Error>> {
     // Parse arguments
     let peptides: Vec<String> = serde_json::from_str(&peptides)?;
-    let taxon_query_ids: Vec<i32> = serde_json::from_str(&taxon_query)?;
+    let taxon_query_ids: Vec<usize> = serde_json::from_str(&taxon_query)?;
     
     // First we retrieve all taxa associated with the given peptids
-    let mut peptides_taxa: HashMap<String, Vec<i32>> = get_taxa_for_peptides(peptides)?;
+    let mut peptides_taxa: HashMap<String, Vec<usize>> = get_taxa_for_peptides(peptides)?;
 
     // Then, we make sure to filter the taxa and only keep those that are associated 
     // to the taxa of interest indicated by the user. Retrieve all (in)direct children
     // of the filter taxa provided by the user
-    let taxa_filter: HashSet<i32> = get_descendants_for_taxa(taxon_query_ids, rank)?;
+    let taxa_filter: HashSet<usize> = get_descendants_for_taxa(taxon_query_ids, rank)?;
 
     // Compute the intersection of the taxa that should be retained and the original list of taxa
     for taxa_list in peptides_taxa.values_mut() {
